Replace validation rules for Product Attribute Controller
For the quantity field in the store() Product Attribute Controller
method, the 'integer' validation rule is measured to 'numeric' as for
a floating-point number.

+++ app/Http/Controllers/Product/ProductAttributeController.php
@@ -34,7 +34,7 @@ class ProductAttributeController extends BaseController
         };
 
         $validatedData = $request->validate([
-            'value' => ['required', 'integer', 'max:255'],
+            'value' => ['required', 'numeric', 'max:255'],
         ]);
 
         $product->attributes()->attach($attribute, ['value' => $validatedData['value']]);
